notes/interface mod
